+++ front-end/src/components/Access.jsx
@@ -1,4 +1,5 @@
 import './Access.css';
+import { useEffect } from 'react';
 
 function Access({access, 
     setAccess, 
@@ -7,31 +8,92 @@ function Access({access,
     password, 
     setPassword,
     confirmPassword,
-    setConfirm
+    setConfirm,
+    parentEmail,
+    setParentEmail,
+    pronouns,
+    setPronouns,
+    birthday,
+    setBirthday,
+    avatar,
+    setAvatar
 }) {
 
-    const handleClick = () => {
-        setAccess(state => !state)
-        setUsername('')
-        setPassword('')
-        setConfirm('')
+    const handleClick = (click) => {
+        if (click.target.id === 'access-return-login' && access == false) {
+            setAccess(state => !state)
+            setUsername('')
+            setPassword('')
+            setConfirm('')
+        } else if (click.target.id === 'access-login' && access === true) {
+            console.log('you have logged in') 
+        } else if (click.target.id === 'access-sign-up' && access === true) {
+            setAccess(state => !state)
+            setUsername('')
+            setPassword('')
+            setConfirm('')
+        } else if (click.target.id === 'access-sign-up' && access === false) {
+            console.log('you have signed up') 
+        }
     }
 
     const handleInfo = (input) => {
-        if (input.target.id === 'username') {
-            setUsername(input.target.value)
-        } else if (input.target.id === 'password') {
-            setPassword(input.target.value)
-        } else if (input.target.id === 'confirm-password') {
-            setConfirm(input.target.value)
+        switch(input.target.id) {
+            case 'username':
+                setUsername(input.target.value)
+                break;
+            case 'password':
+                setPassword(input.target.value)
+                break;
+            case 'confirm-password':
+                setConfirm(input.target.value)
+                break;
+            case 'parent-email':
+                setParentEmail(input.target.value)
+                break;
+            case 'pronouns': 
+                setPronouns(input.target.value)
+                break;
+            default:
+                setUsername('')
+                setPassword('')
+                setConfirm('')
+                setParentEmail('')
+                setPronouns('')
+
         }
     }
+
+    const chooseAvi = (avi) => {
+        switch(avi.target.id) {
+            case 'av-1':
+                setAvatar(avi.target.id)
+                break;
+            case 'av-2':
+                setAvatar(avi.target.id)
+                break;
+            case 'av-3':
+                setAvatar(avi.target.id)
+                break;
+            case 'av-4':
+                setAvatar(avi.target.id)
+                break;
+            case 'av-5':
+                setAvatar(avi.target.id)
+                break;
+            case 'av-6':
+                setAvatar(avi.target.id)
+                break;
+        default:
+                setAvatar(avi.target.id)
+        }
+    }
+
 
     return (
         <div className='access'>
         {access ? 
             <div className='login-div'>
-            <h2>Login Page</h2>
             <form>
 
             <input 
@@ -54,14 +116,15 @@ function Access({access,
 
             : 
             <div className='signup-div'>
-            <h2>Sign Up page</h2>
 
+            <form>
+            <div className='enter-info'>
             <input 
             onChange={handleInfo}
             htmlFor='username'
             id='username'
             type='text'
-            placeholder='username'
+            placeholder='create username'
             value={ username }/>
 
             <input
@@ -69,7 +132,7 @@ function Access({access,
             htmlFor='password'
             id='password'
             type='password'
-            placeholder='password'
+            placeholder='create password'
             value={ password }/>
 
             <input
@@ -80,13 +143,81 @@ function Access({access,
             placeholder='confirm password'
             value={ confirmPassword }/>
 
+            <input
+            onChange={handleInfo}
+            htmlFor='parent email'
+            id='parent-email'
+            type='text'
+            placeholder='parent email'
+            value={parentEmail}/>
+
+            <input
+            onChange={handleInfo}
+            htmlFor='pronouns'
+            id='pronouns'
+            type='text'
+            placeholder='pronouns'
+            value={pronouns}/>
+
+            <input
+            onChange={handleInfo}
+            htmlFor='birthday'
+            id='birthday'
+            type='date'
+            placeholder='confirm password'
+            />
+
+            </div>
+
+            <div className='bio-avatar'>
+            <div id='av-1' className='avatar' onClick={chooseAvi}></div>
+            <div id='av-2' className='avatar' onClick={chooseAvi}></div>
+            <div id='av-3' className='avatar' onClick={chooseAvi}></div>
+            <div id='av-4' className='avatar' onClick={chooseAvi}></div>
+            <div id='av-5' className='avatar' onClick={chooseAvi}></div>
+            <div id='av-6' className='avatar' onClick={chooseAvi}></div>
+            <textarea 
+            placeholder='Please tell us about yourself! :)'
+            rows='5'
+            cols='30'
+            >
+            </textarea>
+            </div>
+            </form>
+
             </div>
         }
-        <button
-        onClick={handleClick}
-        >
-        Change State
-        </button>
+
+        <div className='access-btns'>
+        { access === true ?
+                <input
+            id='log-in-submit'
+            type='submit'
+            value='Login'
+                />
+                : 
+                <button 
+            onClick={handleClick}
+            id='access-return-login'
+                > Return to Login 
+                </button>
+        }
+
+        { access === true ?
+                <button
+            onClick={handleClick}
+            id='access-sign-up'
+                >
+                Sign Up
+                </button>
+                :
+                <input 
+            id='sign-up-submit'
+            type='submit'
+            value='submit'
+               />
+        }
+        </div>
         </div>
 
     )
